IO package cleanup
