+++ launcher/src/main/kotlin/org/sourcegrade/jagr/launcher/io/OutputSerializationScopeImpl.kt
@@ -23,6 +23,18 @@ import com.google.common.io.ByteArrayDataOutput
 import org.sourcegrade.jagr.launcher.env.Jagr
 import org.sourcegrade.jagr.launcher.env.serializerFactoryLocator
 
+fun createScope(output: ByteArrayDataOutput, jagr: Jagr, parent: SerializationScope.Output? = null): SerializationScope.Output {
+  return OutputSerializationScopeImpl(output, jagr, parent)
+}
+
+inline fun <T> openScope(output: ByteArrayDataOutput, jagr: Jagr, block: SerializationScope.Output.() -> T): T {
+  return createScope(output, jagr).block()
+}
+
+inline fun <T> SerializationScope.Output.openScope(block: SerializationScope.Output.() -> T): T {
+  return createScope(output, jagr, this).block()
+}
+
 private class OutputSerializationScopeImpl(
   override val output: ByteArrayDataOutput,
   override val jagr: Jagr,
@@ -32,18 +44,4 @@ private class OutputSerializationScopeImpl(
     SerializerFactory[key.type, jagr.serializerFactoryLocator].write(obj, this)
     backing[key] = obj
   }
-}
-
-fun createScope(output: ByteArrayDataOutput, jagr: Jagr, parent: SerializationScope.Output? = null): SerializationScope.Output {
-  return OutputSerializationScopeImpl(output, jagr, parent)
-}
-
-// fun 1
-inline fun <T> openScope(output: ByteArrayDataOutput, jagr: Jagr, block: SerializationScope.Output.() -> T): T {
-  return createScope(output, jagr).block()
-}
-
-// fun 2
-inline fun <T> SerializationScope.Output.openScope(block: SerializationScope.Output.() -> T): T {
-  return createScope(output, jagr, this).block()
 }
